bcm2835-v4l2-codec: fix vchiq mmal renable
:Release Notes:
Enable vchiq mmal component on stream start after stream stop

:Detailed Notes:
Enable vchiq mmal component on stream start after stream stop
When we stop streaming for flush from gst vrl2 decoder element
then vchiq mmal component is disabled.
But it does not start back when we start the stream back from
gst v4l2 decoder element because the ctx->component_enabled flag
is not set to false while disabling the component. So setting
the flag to false while disabling vchiq mmal component
Also fixes an error in vidioc_s_selection that is using height
instead of crop_height. Might be a typo it is effecting in
dynamic resolution change of the stream.

:Testing Performed:
Tested the 1080p video plays with 22fps in raspberrypi4

:QA Notes:

:Issues Addressed:
[PLAT-94493] Streaming issues in bcm2835-v4l2-codec driver